requirements.txt
# python generics
python-dotenv==1.0.0
chardet==5.2.0
tiktoken==0.5.2
requests==2.31.0
beautifulsoup4==4.12.2
unstructured==0.11.8
colorama==0.4.6
markdownify==0.11.6
pandas==2.2.0
tabulate==0.9.0
python-docx==0.8.11
openpyxl==3.1.2
xlsxwriter==3.1.9

# Open AI
openai==1.11.0

# LangChain
langchain==0.1.8
langchain-community==0.0.21
langchain-experimental==0.0.52
langchain-openai==0.0.5
langgraph==0.0.24
langchainhub==0.1.14
langchain-text-splitters==0.0.1

# Streamlit
streamlit==1.32.1
streamlit-authenticator==0.3.1
streamlit-aggrid==0.3.4.post3
streamlit-autorefresh==1.0.1
extra-streamlit-components==0.1.71

# azure
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-search-documents==11.4.0